editoast: move RollingStock to editoast_models
