+++ editoast/editoast_models/src/rolling_stock/power_restrictions.rs
@@ -1,6 +1,7 @@
 use std::ops::DerefMut;
 
 use database::DbConnection;
+use diesel::QueryableByName;
 use diesel::sql_query;
 use diesel::sql_types::Text;
 use diesel_async::RunQueryDsl;
