+++ editoast/editoast_models/src/rolling_stock/sql/get_power_restrictions.sql
