+++ editoast/src/client/import_rolling_stock.rs
@@ -7,9 +7,9 @@ use clap::Args;
 use colored::Colorize as _;
 use database::DbConnectionPoolV2;
 
-use crate::models::RollingStock;
 use editoast_models::TowedRollingStock;
 use editoast_models::prelude::*;
+use editoast_models::rolling_stock::RollingStock;
 
 #[derive(Args, Clone, Debug)]
 #[command(about, long_about = "Import a rolling stock given a json file")]
