+++ editoast/src/models/fixtures.rs
@@ -60,7 +60,6 @@ use uuid::Uuid;
 use crate::infra_cache::operation::create::apply_create_operation;
 use crate::models;
 use crate::models::Infra;
-use crate::models::RollingStock;
 use crate::models::rolling_stock_livery::RollingStockLivery;
 use editoast_models::Document;
 use editoast_models::ElectricalProfileSet;
@@ -69,6 +68,7 @@ use editoast_models::WorkSchedule;
 use editoast_models::WorkScheduleGroup;
 use editoast_models::prelude::*;
 use editoast_models::project::Project;
+use editoast_models::rolling_stock::RollingStock;
 use editoast_models::scenario::Scenario;
 use editoast_models::study::Study;
 use editoast_models::tags::Tags;
