+++ editoast/src/models/mod.rs
@@ -12,7 +12,6 @@ pub use auth_driver::PgAuthDriver;
 pub mod catalogue_entry;
 pub mod paced_train;
 pub mod railjson;
-pub mod rolling_stock;
 pub mod rolling_stock_livery;
 pub mod stdcm_search_environment;
 pub mod train_schedule_set;
@@ -20,7 +19,6 @@ pub mod train_schedule_set;
 pub use infra::Infra;
 pub use infra_objects::*;
 pub use paced_train::PacedTrain;
-pub use rolling_stock::RollingStock;
 
 #[cfg(test)]
 mod tests {
