+++ editoast/src/views/timetable/stdcm.rs
@@ -38,7 +38,6 @@ use crate::AppState;
 use crate::error::InternalError;
 use crate::error::Result;
 use crate::models::Infra;
-use crate::models::RollingStock;
 use crate::views::AuthenticationExt;
 use crate::views::AuthorizationError;
 use crate::views::timetable::PhysicsConsistParameters;
@@ -47,6 +46,7 @@ use crate::views::timetable::simulation::SimulationResponseSuccess;
 use crate::views::timetable::simulation::consist_train_simulation_batch;
 use editoast_models::WorkSchedule;
 use editoast_models::prelude::*;
+use editoast_models::rolling_stock::RollingStock;
 use editoast_models::timetable::Timetable;
 use editoast_models::train_schedule::TrainSchedule;
